[Update] Added design on navbar
