+++ Web/tues_pairs/public/index.html
@@ -27,8 +27,7 @@
     -->
     <title>React App</title>
   </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+  <body style="margin: 0; background-color: rgb(58, 64, 78);">
     <div id="root"></div>
     <!--
       This HTML file is a template.
