+++ Web/tues_pairs/src/components/App/app.scss
@@ -1,0 +1,3 @@
+.main{
+    display: block;
+}
