+++ Web/tues_pairs/src/components/Pages/Authentication/changeHandler.scss
@@ -2,7 +2,7 @@
 @import url("https://fonts.googleapis.com/css?family=Open+Sans&display=swap");
 
 .ChangeHandler {
-  margin-top: 4em;
+  margin-top: 6rem;
   text-align: center;
   display: flex;
   justify-content: center;
@@ -28,7 +28,7 @@
 }
 
 .ChangeHandler-link {
-  color: orange;
+  color: rgb(252, 152, 0);
 }
 
 @keyframes ChangeHandler-logo-spin {
@@ -42,8 +42,8 @@
 
 //Login Main Container
 .login {
-  width: 27em;
-  height: 40em;
+  width: 20em;
+  height: 30em;
   //min-height: 44em;
   display: flex;
   justify-content: center;
@@ -55,7 +55,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background-color: rgba(59, 64, 78, 1);
+    background-color: rgb(34, 34, 43);
     box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.2);
     border-radius: 4px;
     position: relative;
@@ -71,7 +71,7 @@
     flex-direction: column;
     justify-content: center;
     height: 90%;
-    background-color: orange;
+    background-color: rgb(252, 152, 0);
     width: 100%;
     position: absolute;
     right: -34%;
@@ -100,8 +100,8 @@
       font-size: 21px;
       font-weight: 500;
       color: #fff;
-      margin-right: 3em;
-      margin-left: 3em;
+      margin-right: 2em;
+      margin-left: 2em;
     }
   }
 }
@@ -111,13 +111,13 @@
   font-size: 21px;
   padding: 5px 20px;
   border: 0;
-  background-color: orange;
+  background-color: rgb(252, 152, 0);
   color: #fff;
   border-radius: 3px;
   transition: all 250ms ease-in-out;
   cursor: pointer;
   &:hover {
-    background-color: darkorange;
+    background-color: orange;
   }
   &:focus {
     outline: none;
