+++ Web/tues_pairs/src/components/Pages/Navigation/style.scss
@@ -1,0 +1,40 @@
+.navbar{
+    background-color: rgb(34, 34, 43);
+    font-family: "Open Sans", sans-serif;
+    width: 100%;
+    height: 50%;
+
+    .links{
+        display: flex;
+        justify-content: space-between;
+        
+        img{
+            width: 5%;
+            height: 5%;
+            border-radius: 50%;
+            align-self: center;
+        }
+
+        a{
+            padding: 1em 2em;
+            text-decoration: none;
+            color: white;
+            align-self: center;
+
+            &:hover{
+                background-color: rgb(252, 152, 0);
+            }
+        }
+
+    }
+
+    .logo {
+        display: flex;
+        justify-content: center;
+        img{
+            width: 5%;
+            height: 5%;
+
+        }
+    }
+}
