+++ Web/tues_pairs/src/components/Pages/SignOut/signout.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { withFirebase } from '../../Firebase';
 
 const SignOutButton = ({ firebase }) => (
-    <button type="button" onClick={firebase.doSignOut}>
+    <button classnName="signout-button" type="button" onClick={firebase.doSignOut}>
         Sign Out
     </button>
 );
